Safari-609.1.11.1/LayoutTests/http/tests/security/mixedContent/secure-redirect-to-insecure-redirect-to-basic-auth-secure-image-allowCrossOriginSubresourcesToAskForCredentials.https-expected.txt

CONSOLE MESSAGE: The page at https://127.0.0.1:8443/security/mixedContent/secure-redirect-to-insecure-redirect-to-basic-auth-secure-image-allowCrossOriginSubresourcesToAskForCredentials.https.html was allowed to display insecure content from http://127.0.0.1:8080/resources/redirect.php?url=https://localhost:8443/security/mixedContent/resources/subresource/protected-image.php.

https://127.0.0.1:8443/resources/redirect.php?url=http%3A//127.0.0.1%3A8080/resources/redirect.php%3Furl%3Dhttps%3A//localhost%3A8443/security/mixedContent/resources/subresource/protected-image.php - didReceiveAuthenticationChallenge - Responding with testUser:testPassword
This test loads a secure image that redirects to an insecure image that redirects to a secure image guarded by basic authentication. The image should load.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS did load image.
PASS successfullyParsed is true

TEST COMPLETE
